Reorganizing folder structure of properties panel
