+++ client/src/components/properties-panel/propertiesPanelApplicationDropdown/PropertiesPanelApplicationDropdown.jsx
