+++ client/src/components/properties-panel/propertiesView/PropertiesView.css
